[NFC][SSAF][EntityPointerLevel] Move EntityID-to-EPL map serialization to the EPL module
Factor out the serialization of `std::map<EntityId, EntityPointerLevelSet>`
to `EntityPointerLevelFormat.h`.

+++ clang/lib/ScalableStaticAnalysisFramework/Analyses/PointerFlow/PointerFlowAnalysis.cpp
@@ -54,7 +54,8 @@ Expected<std::unique_ptr<AnalysisResult>> deserializePointerFlowAnalysisResult(
 
   if (Content->size() % 2 != 0)
     return makeSawButExpectedError(
-        *Content, "an even number of elements, got %zu", Content->size());
+        *Content, "an even number of elements, got %lu",
+        static_cast<unsigned long>(Content->size()));
 
   std::map<EntityId, EdgeSet> Edges;
